Switch between journey planner views depending on device type.
