+++ LiveTramsMCR/Shared/Views/Journeys/JourneyPlanViewPad.swift
@@ -1,0 +1,16 @@
+//
+//  JourneyPlanViewPad.swift
+//  LiveTramsMCR (iOS)
+//
+//  Created by David Cook on 07/05/2023.
+//
+
+import Foundation
+import SwiftUI
+
+
+struct JourneyPlanViewPad: View {
+    var body: some View {
+        Text("Placeholder")
+    }
+}
